refactor(sonar): Move localization API to new controller
Extract GetTranslations and related localization logic out of ThemerrController into a new ThemerrLocalizationController. Update ThemerrController to remove the server configuration dependency and unused usings. Update unit tests to instantiate and exercise ThemerrLocalizationController and adjust mocked loggers/configuration accordingly. This separates localization responsibilities and keeps ThemerrController focused on theming operations.
